Kerning is specified as an integer.

+++ python/capypdf.py
@@ -345,7 +345,7 @@ cfunc_types = (
 
 ('capy_text_sequence_new', [ctypes.c_void_p]),
 ('capy_text_sequence_append_codepoint', [ctypes.c_void_p, ctypes.c_uint32]),
-('capy_text_sequence_append_kerning', [ctypes.c_void_p, ctypes.c_double]),
+('capy_text_sequence_append_kerning', [ctypes.c_void_p, ctypes.c_int32]),
 ('capy_text_sequence_append_actualtext_start', [ctypes.c_void_p, ctypes.c_char_p]),
 ('capy_text_sequence_append_actualtext_end', [ctypes.c_void_p]),
 ('capy_text_sequence_append_raw_glyph', [ctypes.c_void_p, ctypes.c_uint32, ctypes.c_uint32]),
